+++ src/compose/add-binding-methods.js
@@ -23,7 +23,11 @@ export default store => {
   store.subscribe(() => {
     const newValues = watchedSelectors.all
       ? store.selectAll()
-      : store.select(Object.keys(watchedSelectors))
+      : // allowMissing is always true here because we don't need to worry about
+        // the selector being missing on every single change, we can just ignore the
+        // missing selectors. Instead, the check for missing selectors (if we care)
+        // is done in the subscribeToSelectors method when we setup the subscription initially
+        store.select(Object.keys(watchedSelectors), { allowMissing: true })
     const { watchedValues } = store.subscriptions
 
     // the only diffing in the app happens here
@@ -67,7 +71,11 @@ export default store => {
   // given an array of selector names, it will call the
   // callback any time those have changed with an object
   // containing only changed values
-  store.subscribeToSelectors = (keys, callback) => {
+  store.subscribeToSelectors = (keys, callback, options) => {
+    // If the selectors don't exist yet they will simply be ignored
+    // this allows a later integratedBundle to add the selector
+    const allowMissing = (options && options.allowMissing) || false
+
     const isAll = keys === 'all'
     // re-use loop for double duty
     // extract names, but also ensure
@@ -83,7 +91,7 @@ export default store => {
     // track changes
     Object.assign(
       store.subscriptions.watchedValues,
-      isAll ? store.selectAll() : store.select(keys)
+      isAll ? store.selectAll() : store.select(keys, { allowMissing })
     )
 
     // return function that can be used to unsubscribe